changes to the UI

+++ PottyTrainer.Contracts/IPottyTrainerRepository.cs
@@ -5,8 +5,8 @@ namespace PottyTrainer.Contracts
     public interface IPottyTrainerRepository
     {
         long SaveEvent(PeePooEvent evt);
-        bool DeleteEvent(long id);
-        PeePooEvent GetEvent(long id);
+        bool DeleteEvent(int id);
+        PeePooEvent GetEvent(int id);
         List<PeePooEvent> GetEvents();
